requirements.txt
cloudscraper>=1.2.71
beautifulsoup4>=4.12.2
lxml>=4.9.2
python-dotenv>=1.0.1
requests>=2.31.0
tqdm>=4.66.1
ujson>=5.8.0
playwright>=1.40.0
mysql-connector-python>=8.0.0
